iwpmd: add NO_PORT_MAP support
A soft iwarp driver that uses the host TCP stack via a kernel mode
socket does not need port mapping.  In fact, if the port map daemon,
iwpmd, is running, then iwpmd must not try and create/bind a socket to
the actual port for a soft iwarp connection, since the driver already
has that socket bound.

Yet if the soft iwarp driver wants to interoperate with hard iwarp
devices that -are- using port mapping, then the soft iwarp driver's
mappings still need to be maintained and advertised by the iwpm protocol.

This patch enhances iwpmd to support the new version 4 IWPM netlink
protocol.  It does so while still providing support for kernels with
ABI version 3.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>