rping: Move QP to error in self created QP mode
When the user is working with self created QP, the rdma_disconnect
doesn't issue modify to error on the QP and it's the application's
responsibility to do so.

When running with '-C' option, not doing so will cause the CQ thread to
get stuck waiting for FLUSHED completions in the teardown flow.

Fixes: 603dba984bb3 ("rping: Add option to create QP independently")
Signed-off-by: Majd Dibbiny <majd@mellanox.com>
Reviewed-by: Yossi Itigin <yosefe@mellanox.com>
Signed-off-by: Yishai Hadas <yishaih@mellanox.com>